# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: InterHive API
  description: InterHive - Industry Readiness & Talent Connect Platform API
  version: 1.0.0
  contact:
    name: InterHive Team
    email: support@interhive.in
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.interhive.in/api/v1
    description: Production Server
  - url: https://staging-api.interhive.in/api/v1
    description: Staging Server
  - url: http://localhost:3000/api/v1
    description: Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        timestamp:
          type: string
          format: date-time
...
    Company:
      type: object
      required:
        - companyInfo
        - contact
      properties:
        companyInfo:
          type: object
          properties:
            name:
              type: string
            legalName:
              type: string
            registrationNumber:
              type: string
            industry:
              type: array
              items:
                type: string
            size:
              type: integer
            foundedYear:
              type: integer
            website:
              type: string
            description:
              type: string
            logo:
              type: string
        contact:
          type: object
          properties:
            primaryContact:
              type: object
              properties:
                email:
                  type: string
                phone:
                  type: string
                address:
                  type: object
                  properties:
                    street:
                      type: string
                    city:
                      type: string
                    state:
                      type: string
                    country:
                      type: string
                    zipCode:
                      type: string

    # Assessment
    Assessment:
      type: object
      required:
        - title
        - type
        - questions
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [technical, soft_skills, aptitude, behavioral, coding, project]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        duration:
          type: integer
        passingScore:
          type: integer
        questions:
          type: array
          items:
            $ref: '#/components/schemas/AssessmentQuestion'
    AssessmentQuestion:
      type: object
      required:
        - type
        - text
        - points
      properties:
        type:
          type: string
          enum: [multiple_choice, multiple_select, coding, essay, practical]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        category:
          type: string
        text:
          type: string
        options:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              text:
                type: string
              isCorrect:
                type: boolean
        points:
          type: integer
        codeSnippet:
          type: string

    # Match
...
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            message: Resource not found
            timestamp: '2024-01-01T00:00:00.000Z'
...
paths:
...
  /companies:
    post:
      tags: [Companies]
      summary: Create company profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Company'
      responses:
        201:
          description: Company created

    get:
      tags: [Companies]
      summary: Get all companies
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
        - name: industry
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, verified, active, suspended, inactive]
      responses:
        200:
          description: Companies retrieved
  /companies/{id}:
    get:
      tags: [Companies]
      summary: Get company by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Company retrieved
        404:
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [Companies]
      summary: Update company
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Company'
      responses:
        200:
          description: Company updated
        404:
          $ref: '#/components/responses/NotFoundError'
...
  /assessments:
    get:
      tags: [Assessments]
      summary: Get all assessments
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, active, archived]
      responses:
        200:
          description: Assessments retrieved

    post:
      tags: [Assessments]
      summary: Create assessment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Assessment'
      responses:
        201:
          description: Assessment created
  /assessments/{id}:
    get:
      tags: [Assessments]
      summary: Get assessment by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Assessment retrieved
  /assessments/{id}/start:
    post:
      tags: [Assessments]
      summary: Start assessment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Assessment started
  /assessments/{id}/submit:
    post:
      tags: [Assessments]
      summary: Submit assessment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - resultId
                - answers
              properties:
                resultId:
                  type: string
                answers:
                  type: array
                  items:
                    type: object
                    properties:
                      questionId:
                        type: string
                      answer:
                        type: string
      responses:
        200:
          description: Assessment submitted

  # Matching
...
  /analytics/dashboard:
    get:
      tags: [Analytics]
      summary: Get dashboard analytics
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        200:
          description: Dashboard analytics retrieved
  /analytics/readiness:
    get:
      tags: [Analytics]
      summary: Get readiness analytics
      security:
        - bearerAuth: []
      responses:
        200:
          description: Readiness analytics retrieved

  # Communication
...